Moving credit card and general payment validation to `lib/checkout` so we can use the validation lib for other payment handlers Updated paths and tests
