+++ client/blocks/credit-card-form/index.jsx
@@ -18,7 +18,7 @@ import { forPayments as countriesList } from 'lib/countries-list';
 import FormButton from 'components/forms/form-button';
 import formState from 'lib/form-state';
 import notices from 'notices';
-import { validateCardDetails } from 'lib/credit-card-details';
+import { validatePaymentDetails } from 'lib/checkout';
 import ValidationErrorList from 'notices/validation-error-list';
 import wpcomFactory from 'lib/wp';
 import { AUTO_RENEWAL, MANAGE_PURCHASES } from 'lib/url/support';
@@ -216,7 +216,7 @@ class CreditCardForm extends Component {
 	}
 
 	getValidationErrors() {
-		const validationResult = validateCardDetails( this.getCardDetails() );
+		const validationResult = validatePaymentDetails( this.getCardDetails() );
 
 		// Maps keys from credit card validator to work with formState.
 		return mapKeys( validationResult.errors, ( value, key ) => {
