+++ client/components/upgrades/credit-card-number-input/index.jsx
@@ -9,7 +9,7 @@ import React from 'react';
 /**
  * Internal dependencies
  */
-import { getCreditCardType } from 'lib/credit-card-details';
+import { getCreditCardType } from 'lib/checkout';
 import Input from 'my-sites/domains/components/form/input';
 
 class CreditCardNumberInput extends React.Component {
