+++ client/lib/checkout/index.js
@@ -29,3 +29,5 @@ export function getExitCheckoutUrl( cart, siteSlug ) {
 
 	return siteSlug ? url + siteSlug : url;
 }
+
+export { getCreditCardType, validatePaymentDetails } from './validation';
