+++ client/lib/checkout/validation.js
@@ -12,6 +12,12 @@ import i18n from 'i18n-calypso';
 import { isEbanxEnabledForCountry, isValidCPF } from 'lib/checkout/ebanx';
 import { PAYMENT_PROCESSOR_EBANX_COUNTRIES } from 'lib/checkout/constants';
 
+/**
+ * Returns ebanx validation rule set. Ebanx required fields may vary according to country
+ * See: https://developers.ebanx.com/api-reference/full-api-documentation/ebanx-payment-2/ebanx-payment-guide/guide-create-a-payment/
+ * @param {string} country two-letter country code to determine the required ebanx fields
+ * @returns {object} the ruleset
+ */
 function ebanxFieldRules( country ) {
 	const requiredFields = PAYMENT_PROCESSOR_EBANX_COUNTRIES[ country ].requiredFields || [];
 
@@ -51,6 +57,11 @@ function ebanxFieldRules( country ) {
 	);
 }
 
+/**
+ * Returns the credit card validation rule set
+ * @param {object} additionalFieldRules custom validation rules depending on jurisdiction or other variable
+ * @returns {object} the ruleset
+ */
 function creditCardFieldRules( additionalFieldRules = {} ) {
 	return Object.assign(
 		{
@@ -92,6 +103,23 @@ function creditCardFieldRules( additionalFieldRules = {} ) {
 		},
 		additionalFieldRules
 	);
+}
+
+/**
+ * Returns a validation ruleset to use for the given payment type
+ * @param {object} paymentDetails object containing fieldname/value keypairs
+ * @param {string} paymentType credit-card(default)|paypal|ideal|p24|tef
+ * @returns {object|null} the ruleset
+ */
+function paymentFieldRules( paymentDetails, paymentType ) {
+	switch ( paymentType ) {
+		case 'tef':
+			return ebanxFieldRules( 'BR' );
+		case 'credit-card':
+			return creditCardFieldRules( getAdditionalFieldRules( paymentDetails ) );
+		default:
+			return null;
+	}
 }
 
 function parseExpiration( value ) {
@@ -172,11 +200,20 @@ validators.validCPF = {
 	},
 };
 
-export function validateCardDetails( cardDetails ) {
-	const rules = creditCardFieldRules( getAdditionalFieldRules( cardDetails ) ),
+/**
+ * Runs payment fields through the relevant validation rules
+ * use these validation rules, for example, in <CreditCardForm />, <PayPalPaymentBox /> and <RedirectPaymentBox />
+ * @param {object} paymentDetails object containing fieldname/value keypairs
+ * @param {string} paymentType credit-card(default)|paypal|ideal|p24|tef
+ * @returns {object} validation errors, if any
+ */
+export function validatePaymentDetails( paymentDetails, paymentType = 'credit-card' ) {
+	const rules = paymentFieldRules( paymentDetails, paymentType );
+	let errors = [];
+	if ( rules ) {
 		errors = Object.keys( rules ).reduce( function( allErrors, fieldName ) {
 			const field = rules[ fieldName ],
-				newErrors = getErrors( field, cardDetails[ fieldName ], cardDetails );
+				newErrors = getErrors( field, paymentDetails[ fieldName ], paymentDetails );
 
 			if ( newErrors.length ) {
 				allErrors[ fieldName ] = newErrors;
@@ -184,8 +221,8 @@ export function validateCardDetails( cardDetails ) {
 
 			return allErrors;
 		}, {} );
-
-	return { errors: errors };
+	}
+	return { errors };
 }
 
 /**
@@ -221,12 +258,19 @@ export function getCreditCardType( number ) {
 	return null;
 }
 
-function getErrors( field, value, cardDetails ) {
+/**
+ *
+ * @param {string} field the name of the field
+ * @param {value} value the value of the field
+ * @param {object} paymentDetails object containing fieldname/value keypairs
+ * @returns {array} array of errors found, if any
+ */
+function getErrors( field, value, paymentDetails ) {
 	return compact(
 		field.rules.map( function( rule ) {
 			const validator = getValidator( rule );
 
-			if ( ! validator.isValid( value, cardDetails ) ) {
+			if ( ! validator.isValid( value, paymentDetails ) ) {
 				return validator.error( field.description );
 			}
 		} )
