+++ client/lib/credit-card-details/index.js
@@ -3,4 +3,3 @@
  * Internal dependencies
  */
 export { maskField, unmaskField } from './masking';
-export { getCreditCardType, validateCardDetails } from './validation';
